backend/requirements.txt
Flask>=2.0
Flask-Cors
Flask-SocketIO>=5.3.0
python-dotenv
marshmallow
asgiref
pymongo
bcrypt
gunicorn==20.1.0
PyJWT
werkzeug==2.0.2
black==23.3.0
httpx==0.27.0
aiohttp==3.9.3
redis>=4.5.0
eventlet>=0.33.0
